подключил framework Emotion, отрефакторил тесты и структуру проекта в древовидную
