+++ .eslintrc
@@ -17,6 +17,21 @@
         "endOfLine": "auto"
       }
     ],
+    "no-restricted-syntax": [
+      "error",
+      {
+        "selector": "ExportDefaultDeclaration",
+        "message": "Restricted default export, prefer named exports!"
+      }
+    ],
+    "no-console": [
+      "error",
+      { "allow": [
+          "warn",
+          "error"
+        ]
+      }
+    ],
     "@typescript-eslint/explicit-function-return-type": 0,
     "react/prop-types": 0
   },
@@ -30,6 +45,14 @@
       "files": ["webpack.config.js"],
       "rules": {
         "@typescript-eslint/no-var-requires": ["off"]
+      }
+    },
+    {
+      "files": [
+        "*.stories.tsx"
+      ],
+      "rules": {
+        "no-restricted-syntax": ["off"]
       }
     }
   ]
