+++ package.json
@@ -48,10 +48,13 @@
     "@types/jest": "^25.1.4",
     "@types/react": "^16.9.33",
     "@types/react-dom": "^16.9.6",
+    "@types/react-test-renderer": "^16.9.2",
     "@typescript-eslint/eslint-plugin": "^2.25.0",
     "@typescript-eslint/parser": "^2.25.0",
     "babel-jest": "^25.2.4",
     "babel-loader": "^8.1.0",
+    "babel-plugin-emotion": "^10.0.33",
+    "css-loader": "^3.5.2",
     "enzyme": "^3.11.0",
     "enzyme-adapter-react-16": "^1.15.2",
     "eslint": "^6.8.0",
@@ -65,6 +68,7 @@
     "loki": "^0.20.3",
     "prettier": "^2.0.2",
     "react-docgen-typescript-loader": "^3.7.2",
+    "react-test-renderer": "^16.13.1",
     "storybook-addon-react-docgen": "^1.2.32",
     "ts-node": "^8.8.2",
     "typescript": "^3.8.3",
@@ -73,6 +77,8 @@
     "webpack-dev-server": "^3.10.3"
   },
   "dependencies": {
+    "@emotion/core": "^10.0.28",
+    "@emotion/styled": "^10.0.27",
     "react": "^16.13.1",
     "react-dom": "^16.13.1"
   },
