--- src/calc/helpers.ts
@@ -1,1 +1,0 @@
-export const isNumber = (item: string): boolean => !isNaN(Number(item));
