--- src/calc/parser.test.ts
@@ -1,62 +1,0 @@
-import { parser } from "./parser";
-
-describe("Parser correct cases", () => {
-  it("2 **!", () => {
-    expect(parser("2 **")).toEqual([2, "**"]);
-  });
-
-  it("1 + 32 !", () => {
-    expect(parser("1 + 32 !")).toEqual([1, "+", 32, "!"]);
-  });
-
-  it("1 + 32 ^ 5", () => {
-    expect(parser("1 + 32 ^ 5")).toEqual([1, "+", 32, "^", 5]);
-  });
-
-  it("1 + 32 ^ 5 !", () => {
-    expect(parser("1 + 32 ^ 5 !")).toEqual([1, "+", 32, "^", 5, "!"]);
-  });
-
-  it("1 + 32", () => {
-    expect(parser("1 + 32")).toEqual([1, "+", 32]);
-  });
-
-  it("11 + 3 * 22", () => {
-    expect(parser("11 + 3 * 22")).toEqual([11, "+", 3, "*", 22]);
-  });
-
-  it("1 + 32 - 2 + 2", () => {
-    expect(parser("1 + 32 - 2 + 2")).toEqual([1, "+", 32, "-", 2, "+", 2]);
-  });
-
-  it("1 + 32 - 2 ** + 2", () => {
-    expect(parser("1 + 32 - 2 ** + 2")).toEqual([
-      1,
-      "+",
-      32,
-      "-",
-      2,
-      "**",
-      "+",
-      2,
-    ]);
-  });
-});
-
-describe("Parser invalid cases", () => {
-  it("1 + + 33 - 2", () => {
-    expect(() => parser("1 + + 33 - 2")).toThrow(
-      TypeError("Unexpected string")
-    );
-  });
-
-  it("1 ! 33 - 2", () => {
-    expect(() => parser("1 ! 33 - 2")).toThrow(TypeError("Unexpected string"));
-  });
-
-  it("1 ! + + 33 - 2", () => {
-    expect(() => parser("1 ! + + 33 - 2")).toThrow(
-      TypeError("Unexpected string")
-    );
-  });
-});
