--- src/calc/runner.ts
@@ -1,28 +1,0 @@
-import { parser } from "./parser";
-
-import {
-  firstPrioritiesCalc,
-  secondPrioritiesCalc,
-  thirdPrioritiesCalc,
-} from "./engine";
-
-export const runner = (line: string): number => {
-  const stack = parser(line);
-
-  if (stack === null) {
-    throw new TypeError("Unexpected string");
-  }
-
-  const firstPrioritiesRes = firstPrioritiesCalc(stack);
-
-  if (firstPrioritiesRes.length === 1) {
-    return Number(firstPrioritiesRes[0]);
-  }
-
-  const secondPrioritiesRes = secondPrioritiesCalc(firstPrioritiesRes);
-  if (secondPrioritiesRes.length === 1) {
-    return Number(secondPrioritiesRes[0]);
-  }
-
-  return thirdPrioritiesCalc(secondPrioritiesRes);
-};
