--- src/index.html
@@ -1,10 +1,0 @@
-<!DOCTYPE html>
-<html lang="ru">
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>React project</title>
-</head>
-<body>
-</body>
-</html>
