--- src/lesson4.5/generics.ts
@@ -1,6 +1,0 @@
-// Playground for ArrayLike generic illustration
-const divNodeList = document.querySelectorAll("div");
-
-const divArray = Array.from(divNodeList);
-
-divArray.pop();
