--- src/lesson4.5/homework/README.md
@@ -1,8 +1,0 @@
-Задание к TypeScript: Part 2 будет проверять ваше умение использовать и вычислять типы
-Все задания устроены таким образом что в них есть тип FIXME (который any) и ваша задача избавится от него
-Менять код кроме типов нельзя, исходные типы менять тоже нельзя, но можно рефакторить
-Например `type A = 1 | 2` выразить как `type A1 = 1; type A2 = 2; type A = A1 | A2;`
-
-Для принятия дз достачно сделать первый уровень
-Второй уровень для пытливых
-И есть супер челлендж (hard) для самых упорных
