--- src/lesson4.5/homework/easy2.ts
@@ -1,44 +1,0 @@
-// Задание первого уровня 2
-// Есть объединение (юнион) типов заказов в различных состояниях
-// Нужно заменить FIXME на тип который достанет из Order все возможные состояния (state)
-
-type FIXME =
-  | "initial"
-  | "inWork"
-  | "buyingSupplies"
-  | "producing"
-  | "fullfilled";
-
-type Order =
-  | {
-      state: "initial";
-      sum: number;
-    }
-  | {
-      state: "inWork";
-      sum: number;
-      workerId: number;
-    }
-  | {
-      state: "buyingSupplies";
-      sum: number;
-      workerId: number;
-      suppliesSum: number;
-    }
-  | {
-      state: "producing";
-      sum: number;
-      workerId: number;
-      suppliesSum: number;
-      produceEstimate: Date;
-    }
-  | {
-      state: "fullfilled";
-      sum: number;
-      workerId: number;
-      suppliesSum: number;
-      produceEstimate: Date;
-      fullfillmentDate: Date;
-    };
-
-export const getOrderState = (order: Order): FIXME => order.state;
