--- src/lesson4.5/homework/easy3.ts
@@ -1,15 +1,0 @@
-// Задание первого уровня 3
-// Есть общая функция omit которая удаляет поле из объекта и возвращает его без этого поля
-// Нужно заменить FIXME на соответствующий тип
-
-type FIXME = {};
-
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-export const omit = <T extends Record<any, any>, K extends keyof T>(
-  obj: T,
-  keyToOmit: K
-): FIXME => {
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  const { [keyToOmit]: _, ...withoutKey } = obj;
-  return withoutKey;
-};
