--- src/lesson4.5/homework/medium1.ts
@@ -1,57 +1,0 @@
-// Задание второго уровня 1
-// Есть объединение (юнион) типов заказов в различных состояниях
-// и функция filterOnlyInitialAndInWorkOrder которая принимает заказы в любых состояниях
-// А возвращает только initial и inWork
-// Нужно заменить FIXME на правильный тип вычисленный на основе Order
-
-type FIXME =
-  | {
-      state: "initial";
-      sum: number;
-    }
-  | {
-      state: "inWork";
-      sum: number;
-      workerId: number;
-    }
-  | null;
-
-type Order =
-  | {
-      state: "initial";
-      sum: number;
-    }
-  | {
-      state: "inWork";
-      sum: number;
-      workerId: number;
-    }
-  | {
-      state: "buyingSupplies";
-      sum: number;
-      workerId: number;
-      suppliesSum: number;
-    }
-  | {
-      state: "producing";
-      sum: number;
-      workerId: number;
-      suppliesSum: number;
-      produceEstimate: Date;
-    }
-  | {
-      state: "fullfilled";
-      sum: number;
-      workerId: number;
-      suppliesSum: number;
-      produceEstimate: Date;
-      fullfillmentDate: Date;
-    };
-
-export const filterOnlyInitialAndInWorkOrder = (order: Order): FIXME => {
-  if (order.state === "initial" || order.state === "inWork") {
-    return order;
-  }
-
-  return null;
-};
