--- src/lesson4.5/homework/medium3.ts
@@ -1,18 +1,0 @@
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-type FIXME = string[];
-
-const orderStates = [
-  "initial",
-  "inWork",
-  "buyingSupplies",
-  "producing",
-  "fullfilled",
-] as const;
-
-type OrderState = typeof orderStates[number];
-
-// Hint: type guards
-export const getUserOrderStates = (orderStates: OrderState[]): FIXME =>
-  orderStates.filter(
-    (state) => state !== "buyingSupplies" && state !== "producing"
-  );
