--- src/lesson4/Field.css
@@ -1,5 +1,0 @@
-.field {
-  display: inline-block;
-  padding: 10px;
-  border: 2px solid lightgray;
-}
