+++ src/lesson5/components/InteractiveField/InteractiveField.stories.tsx
@@ -1,0 +1,18 @@
+import React from "react";
+import { withKnobs, number } from "@storybook/addon-knobs";
+import { Field } from "./components";
+import { InteractiveField } from "./InteractiveField";
+export default {
+  title: "InteractiveField",
+  decorators: [withKnobs],
+};
+
+const players = " ";
+export const withRealField = () => (
+  <InteractiveField
+    xSize={number("xSize", 10)}
+    ySize={number("ySize", 10)}
+    playerMarks={players}
+    fieldComponent={Field}
+  />
+);
