+++ src/lesson5/components/InteractiveField/components/Field/Field.stories.tsx
@@ -1,0 +1,35 @@
+import React from "react";
+import { action } from "@storybook/addon-actions";
+import { withKnobs, object } from "@storybook/addon-knobs";
+import { Field } from "./Field";
+
+export default {
+  title: "FieldContainer",
+  decorators: [withKnobs],
+};
+
+const elementClickedJsx = action("Cell clicked (jsx)");
+
+export const emptyField = () => [
+  <Field
+    key="jsx"
+    onClick={elementClickedJsx}
+    field={object("field", [
+      ["", "", ""],
+      ["", "", ""],
+      ["", "", ""],
+    ])}
+  />,
+];
+
+export const nonEmptyField = () => [
+  <Field
+    key="jsx"
+    onClick={elementClickedJsx}
+    field={object("field", [
+      [" ", " ", ""],
+      ["", " ", ""],
+      ["", "", ""],
+    ])}
+  />,
+];
