+++ src/lesson5/components/InteractiveField/components/Field/Field.tsx
@@ -1,0 +1,16 @@
+import React, { FC } from "react";
+import { Cell } from "./components";
+import { FieldContainer } from "./FieldContainer";
+
+import type { FieldProps } from "types/field";
+
+export const Field: FC<FieldProps> = ({ field, onClick }) => (
+  <FieldContainer>
+    {field.map((row, y) => [
+      ...row.map((filled: string, x) => (
+        <Cell key={`${x}_${y}`} filled={filled} x={x} y={y} onClick={onClick} />
+      )),
+      y !== row.length - 1 ? <br key={y} /> : null,
+    ])}
+  </FieldContainer>
+);
