+++ src/lesson5/components/InteractiveField/components/Field/FieldContainer.tsx
@@ -1,0 +1,12 @@
+import styled from "@emotion/styled";
+import { css } from "@emotion/core";
+
+const Field = css`
+  display: inline-block;
+  padding: 10px;
+  border: 2px solid lightgray;
+`;
+
+export const FieldContainer = styled.div`
+  ${Field};
+`;
