+++ src/lesson5/components/InteractiveField/components/Field/components/Cell/Cell.tsx
@@ -1,0 +1,13 @@
+import React, { FC } from "react";
+import type { CellProps } from "types/field";
+
+import { CellItem } from "./CellItems";
+
+export const Cell: FC<CellProps> = ({ filled, x, y, onClick }) => {
+  const isFilled = Boolean(filled);
+  return (
+    <CellItem isFilled={isFilled} onClick={() => onClick(x || 0, y || 0)}>
+      {filled}
+    </CellItem>
+  );
+};
