+++ src/lesson5/components/InteractiveField/components/Field/components/Cell/index.ts
@@ -1,0 +1,1 @@
+export * from "./Cell";
