+++ src/lesson5/components/InteractiveField/components/Field/components/index.ts
@@ -1,0 +1,1 @@
+export * from "./Cell";
