+++ src/lesson5/components/InteractiveField/components/index.ts
@@ -1,0 +1,1 @@
+export * from "./Field";
