+++ src/lesson5/components/InteractiveField/index.ts
@@ -1,0 +1,2 @@
+export * from "./InteractiveField";
+export * from "./components";
