+++ tsconfig.json
@@ -1,5 +1,9 @@
 {
   "compilerOptions": {
+    "baseUrl": "./",
+    "paths": {
+      "types/*": ["src/lesson5/types/*"]
+    },
     "target": "es5",
     "module": "commonjs",
     "allowJs": true,
